backend/auth-service/src/requirements.txt
redis==5.0.1
fastapi==0.100.1
pydantic==2.4.2
pydantic_settings==2.0.3
uvicorn==0.24.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
gunicorn==21.2.0
httptools==0.6.1
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
async_fastapi_jwt_auth==0.6.1
passlib==1.7.4
typer==0.9.0

werkzeug==3.0.1

pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.8.6
backoff==2.2.1
requests==2.31.0

opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

Authlib==1.2.1
itsdangerous==2.1.2
